Move test utils to different package
