+++ modules/ballerina-test/src/test/java/org/ballerinalang/test/utils/debug/DebuggerExecutor.java
@@ -15,7 +15,7 @@
 *  specific language governing permissions and limitations
 *  under the License.
 */
-package org.ballerinalang.test.debugger;
+package org.ballerinalang.test.utils.debug;
 
 import org.ballerinalang.BLangProgramRunner;
 import org.ballerinalang.bre.Context;
