+++ modules/ballerina-test/src/test/java/org/ballerinalang/test/utils/debug/TestDebugServer.java
@@ -15,7 +15,7 @@
 * specific language governing permissions and limitations
 * under the License.
 */
-package org.ballerinalang.test.debugger;
+package org.ballerinalang.test.utils.debug;
 
 import org.ballerinalang.util.debugger.DebugServer;
 import org.ballerinalang.util.debugger.VMDebugManager;
